cityCategory taken into single state, emptied App.js file
