+++ client/src/HambergerApp.css
@@ -1,0 +1,3 @@
+.HambergerApp {
+  text-align: center;
+}
