+++ client/src/HambergerApp.test.js
@@ -1,8 +1,8 @@
 import { render, screen } from "@testing-library/react";
-import App from "./App";
+import HambergerApp from "./HambergerApp";
 
 test("renders learn react link", () => {
-  render(<App />);
+  render(<HambergerApp />);
   const linkElement = screen.getByText(/learn react/i);
   expect(linkElement).toBeInTheDocument();
 });
